refactor table block components into separate files
