+++ apps/web/modules/components/editor/blocks/sdk/table.ts
@@ -3,7 +3,7 @@ import { SYSTEM_IDS } from '@geogenesis/ids';
 import { Entity } from '~/modules/entity';
 import { ID } from '~/modules/id';
 import { Triple } from '~/modules/triple';
-import { Column, EntityValue, Entity as IEntity, Triple as ITriple, TripleValueType } from '~/modules/types';
+import { EntityValue, Entity as IEntity, Triple as ITriple, TripleValueType } from '~/modules/types';
 
 export function upsertName({
   blockEntity,
