+++ apps/web/modules/hooks/use-user-can-edit.ts
@@ -1,0 +1,9 @@
+import { useAccessControl } from '../auth/use-access-control';
+import { useEditable } from '../stores/use-editable';
+
+export function useUserCanEdit(spaceId?: string) {
+  const { editable } = useEditable();
+  const { isEditor } = useAccessControl(spaceId);
+
+  return editable && isEditor;
+}
